Rename AssetGroupCard to AssetGroupPanel.
